requirements.txt
nltk==3.8.1
unidecode==1.3.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
#tkinter  # Viene incluido con Python, no necesita versión específica 